+++ src/relay/CHANGELOG.md
@@ -3,6 +3,19 @@
 - **Breaking**: Relay version upgraded to 12.0.0, for more information please visit: https://github.com/facebook/relay/releases/tag/v12.0.0
 - **Breaking**: All previous loggers (`RelayEagerLogger`, `RelayLazyLogger` and `RelayDebugLogger`) were replaced with one simple `RelayLogger`. This allows us to focus better on one solution that is more friendly with the new Relay Hooks development workflow.
 - **Breaking**: function `commitMutationAsync` has been removed. You can still use legacy `commitMutation` or preferably `useMutation` hook which is much better replacement for these functions.
+- **Breaking**: types `PaginationRelayProp` and `RefetchRelayProp` were renamed to `RelayPaginationProp` and `RelayRefetchProp` respectively. Additionally, they now expect component instead of props type in the generic type (that applies to `RelayProp` type as well). Migration example:
+
+  ```diff
+  - }): FragmentContainerType<Props>);
+  + }): RelayFragmentContainer<typeof DefaultComponent>);
+
+  - ): PaginationContainerType<Props>);
+  + ): RelayPaginationContainer<typeof DefaultComponent>);
+
+  - ): RefetchContainerType<Props>);
+  + ): RelayRefetchContainer<typeof DefaultComponent>);
+  ```
+
 - **Breaking**: new version of `fetchQuery` from Relay 11 (returns observable instead of promise). For more information please visit: https://github.com/facebook/relay/releases/tag/v11.0.0. Easiest migration path (BUT be aware of [the limitations](https://relay.dev/docs/api-reference/fetch-query/#behavior-with-topromise):
 
   ```diff